move most logic to pkg, keep internal for printing
